refactor: extract parse_jsonl_lines to euroeval.jsonl_io module
- Create src/euroeval/jsonl_io.py with standalone parse_jsonl_lines()
  function extracted from BenchmarkResult.load_from_jsonl()
- Update BenchmarkResult.load_from_jsonl() to use the new function
- Update src/leaderboards/jsonl_io.py to import from euroeval.jsonl_io
- Fix broken imports in scripts (leaderboards.utils → leaderboards.jsonl_io)
- Remove unused json import from data_models.py
